Support building the crate on Windows
This commit makes various changes to the build script so that this crate
can be built on Windows using MSYS2 and similar Unix like environments.
Building using MSVC is not supported at this time, as I could not get
this to work. There are various forks that use CMake, but they are
either based on older versions with the CMake configuration not
supporting newer releases, only support a very limited set of targets,
or don't work for some other reason.