Remove bindgen dependency
This removes the dependency on bindgen, replacing it with a set of
bindings directly included in the repository. The bindings were
generated on Linux, cleaned up, then compared with the bindgen output
for MacOS and Windows (using MSYS2). Apart from some noise caused by
unused constants, the libffi API code was the same.

Removing bindgen cuts down the amount of dependencies, most notably the
dependency on libclang. Removing this simplifies the installation
procedure, as well as simplifying the build scripts a bit.

The complex FFI types are hidden behind a "complex" feature flag that
is disabled by default, similar to the libffi-rs crate. The longdouble
type is disabled on GNU/Linux ARM devices, per
https://github.com/tov/libffi-sys-rs/pull/37#issuecomment-701645235 and
https://github.com/tov/libffi-rs/pull/14.

This fixes https://github.com/tov/libffi-sys-rs/issues/31.